- Linked actions with source positions after loading to simplify ActionEntity, import/export and to avoid data duplication - Rearranged import/export columns of Actions - Removed unused ActionRepository methods - Removed unused methods in ActionEntityDao

+++ app/src/test/java/com/ceaver/assin/positions/PositionAsserts.kt
@@ -13,7 +13,7 @@ fun assertThatPosition(actual: Position): PositionIdAssert {
 
 class PositionIdAssert(actual: Position) : AbstractAssert<PositionIdAssert, Position>(actual, PositionIdAssert::class.java) {
 
-    fun hasId(positionId: Long) {
+    fun hasId(positionId: Int) {
         Assertions.assertThat(actual.id).isEqualTo(positionId)
     }
